Renaming module charts to graphs
